Revert "E2vid pip"
