+++ python/offline_reconstruction.py
@@ -2,17 +2,26 @@ import argparse
 import os
 os.environ['CUDA_DEVICE_ORDER'] = 'PCI_BUS_ID'
 from pathlib import Path
+import urllib
 import warnings
 
+import torch
 import tqdm
 
 from data.provider import DataProvider
 from data.rectimestamps import TimestampProviderFile, TimestampProviderRate
+from reconstruction.utils.loading_utils import load_model, get_device
+from reconstruction.image_reconstructor import ImageReconstructor
+from reconstruction.options.inference_options import set_inference_options
+from reconstruction.utils.voxelgrid import VoxelGrid
 
-import e2vid
-from e2vid.options.inference_options import set_inference_options
-from e2vid.utils.voxelgrid import VoxelGrid
 
+def download_checkpoint(path_to_model):
+    print('Downloading E2VID checkpoint to {} ...'.format(path_to_model))
+    e2vid_model = urllib.request.urlopen('http://rpg.ifi.uzh.ch/data/E2VID/models/E2VID_lightweight.pth.tar')
+    with open(path_to_model, 'w+b') as f:
+        f.write(e2vid_model.read())
+    print('Done with downloading!')
 
 
 if __name__ == "__main__":
@@ -20,12 +29,17 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         description='Image reconstruction')
     parser.add_argument('--h5file', help='Path to h5 file containing events for reconstruction.', default='')
+    parser.add_argument('-c', '--path_to_model', type=str,
+                        help='path to the model weights',
+                        default='reconstruction/pretrained/E2VID_lightweight.pth.tar')
     parser.add_argument('--height', type=int, default=480)
     parser.add_argument('--width', type=int, default=640)
+    parser.add_argument('--gpu_id',  type=int, default=0)
     parser.add_argument('--freq_hz', '-fhz', type=int, default=0, help='Frequency for saving the reconstructed images from events')
     parser.add_argument('--timestamps_file', '-tsf', help='Path to txt file containing image reconstruction timestamps')
     parser.add_argument('--upsample_rate', '-u', type=int, default=1, help='Multiplies the number of reconstructions, which effectively lowers the time window of events for E2VID. These intermediate reconstructions will not be saved to disk.')
-    
+    parser.add_argument('--verbose', '-v',  action='store_true', default=False, help='Verbose output')
+
     set_inference_options(parser)
 
     args = parser.parse_args()
@@ -47,17 +61,24 @@ if __name__ == "__main__":
     data_provider = DataProvider(h5_path, height=args.height, width=args.width, timestamp_provider=timestamp_provider)
 
     # Load model to device
-    reconstructor = e2vid.E2VID(args)
+    if not os.path.isfile(args.path_to_model):
+        download_checkpoint(args.path_to_model)
+    assert os.path.isfile(args.path_to_model)
+    model = load_model(args.path_to_model)
+    device = get_device(args.use_gpu, args.gpu_id)
+    model = model.to(device)
+    model.eval()
 
     if not os.path.exists(args.output_folder):
         os.makedirs(args.output_folder)
     else:
         assert os.path.isdir(args.output_folder)
 
+    image_reconstructor = ImageReconstructor(model, args.height, args.width, model.num_bins, args)
     print('== Image reconstruction == ')
     print('Image size: {}x{}'.format(args.height, args.width))
     print('Will write images to: {}'.format(os.path.join(args.output_folder, args.dataset_name)))
-    grid = VoxelGrid(reconstructor.model.num_bins, args.width, args.height, upsample_rate=args.upsample_rate)
+    grid = VoxelGrid(model.num_bins, args.width, args.height, upsample_rate=args.upsample_rate)
     pbar = tqdm.tqdm(total=len(data_provider))
     for events in data_provider:
         if events.events.size > 0:
@@ -69,8 +90,9 @@ if __name__ == "__main__":
                 else:
                     event_grid, _ = grid.events_to_voxel_grid(sliced_events[i])
                     event_grid = grid.normalize_voxel(event_grid)
-                    reconstructor(event_grid)
+                    event_tensor = torch.from_numpy(event_grid)
+                    image_reconstructor.update_reconstruction(event_tensor)
                 if i== len(sliced_events) - 1:
                     rec_ts_nanoseconds = int(events.t_reconstruction)*1000
-                    reconstructor.image_reconstructor.save_reconstruction(rec_ts_nanoseconds)
+                    image_reconstructor.save_reconstruction(rec_ts_nanoseconds)
                     pbar.update(1)
