--- python/reconstruction/e2vid/options/__init__.py
@@ -1,1 +1,0 @@
-from . import inference_options
