+++ python/reconstruction/utils/event_tensor_utils.py
@@ -1,5 +1,6 @@
 import torch
 import numpy as np
+from .timers import CudaTimer, Timer
 
 
 class EventPreprocessor:
@@ -43,7 +44,8 @@ class EventPreprocessor:
 
         # Remove (i.e. zero out) the hot pixels
         if self.hot_pixel_mask is not None:
-            events = events * self.hot_pixel_mask
+            with CudaTimer('Remove hot pixels'):
+                events = events * self.hot_pixel_mask
 
         # Flip tensor vertically and horizontally
         if self.flip:
@@ -52,17 +54,18 @@ class EventPreprocessor:
         # Normalize the event tensor (voxel grid) so that
         # the mean and stddev of the nonzero values in the tensor are equal to (0.0, 1.0)
         if not self.no_normalize:
-            nonzero_ev = (events != 0)
-            num_nonzeros = nonzero_ev.sum()
-            if num_nonzeros > 0:
-                # compute mean and stddev of the **nonzero** elements of the event tensor
-                # we do not use PyTorch's default mean() and std() functions since it's faster
-                # to compute it by hand than applying those funcs to a masked array
+            with CudaTimer('Normalization'):
+                nonzero_ev = (events != 0)
+                num_nonzeros = nonzero_ev.sum()
+                if num_nonzeros > 0:
+                    # compute mean and stddev of the **nonzero** elements of the event tensor
+                    # we do not use PyTorch's default mean() and std() functions since it's faster
+                    # to compute it by hand than applying those funcs to a masked array
 
-                mean = torch.sum(events, dtype=torch.float32) / num_nonzeros  # force torch.float32 to prevent overflows when using 16-bit precision
-                stddev = torch.sqrt(torch.sum(events ** 2, dtype=torch.float32) / num_nonzeros - mean ** 2)
-                mask = nonzero_ev.type_as(events)
-                events = mask * (events - mean) / stddev
+                    mean = torch.sum(events, dtype=torch.float32) / num_nonzeros  # force torch.float32 to prevent overflows when using 16-bit precision
+                    stddev = torch.sqrt(torch.sum(events ** 2, dtype=torch.float32) / num_nonzeros - mean ** 2)
+                    mask = nonzero_ev.type_as(events)
+                    events = mask * (events - mean) / stddev
 
         return events
 
@@ -127,6 +130,8 @@ def events_to_voxel_grid_pytorch(events, num_bins, width, height, device):
     :return voxel_grid: PyTorch event tensor (on the device specified)
     """
 
+    DeviceTimer = CudaTimer if device.type == 'cuda' else Timer
+
     assert(events.shape[1] == 4)
     assert(num_bins > 0)
     assert(width > 0)
@@ -137,45 +142,47 @@ def events_to_voxel_grid_pytorch(events, num_bins, width, height, device):
     with torch.no_grad():
 
         events_torch = torch.from_numpy(events)
-        events_torch = events_torch.to(device)
+        with DeviceTimer('Events -> Device (voxel grid)'):
+            events_torch = events_torch.to(device)
 
-        voxel_grid = torch.zeros(num_bins, height, width, dtype=torch.float32, device=device).flatten()
+        with DeviceTimer('Voxel grid voting'):
+            voxel_grid = torch.zeros(num_bins, height, width, dtype=torch.float32, device=device).flatten()
 
-        # normalize the event timestamps so that they lie between 0 and num_bins
-        last_stamp = events_torch[-1, 0]
-        first_stamp = events_torch[0, 0]
-        deltaT = last_stamp - first_stamp
+            # normalize the event timestamps so that they lie between 0 and num_bins
+            last_stamp = events_torch[-1, 0]
+            first_stamp = events_torch[0, 0]
+            deltaT = last_stamp - first_stamp
 
-        if deltaT == 0:
-            deltaT = 1.0
+            if deltaT == 0:
+                deltaT = 1.0
 
-        events_torch[:, 0] = (num_bins - 1) * (events_torch[:, 0] - first_stamp) / deltaT
-        ts = events_torch[:, 0]
-        xs = events_torch[:, 1].long()
-        ys = events_torch[:, 2].long()
-        pols = events_torch[:, 3].float()
-        pols[pols == 0] = -1  # polarity should be +1 / -1
+            events_torch[:, 0] = (num_bins - 1) * (events_torch[:, 0] - first_stamp) / deltaT
+            ts = events_torch[:, 0]
+            xs = events_torch[:, 1].long()
+            ys = events_torch[:, 2].long()
+            pols = events_torch[:, 3].float()
+            pols[pols == 0] = -1  # polarity should be +1 / -1
 
-        tis = torch.floor(ts)
-        tis_long = tis.long()
-        dts = ts - tis
-        vals_left = pols * (1.0 - dts.float())
-        vals_right = pols * dts.float()
+            tis = torch.floor(ts)
+            tis_long = tis.long()
+            dts = ts - tis
+            vals_left = pols * (1.0 - dts.float())
+            vals_right = pols * dts.float()
 
-        valid_indices = tis < num_bins
-        valid_indices &= tis >= 0
-        voxel_grid.index_add_(dim=0,
-                                index=xs[valid_indices] + ys[valid_indices] *
-                                width + tis_long[valid_indices] * width * height,
-                                source=vals_left[valid_indices])
+            valid_indices = tis < num_bins
+            valid_indices &= tis >= 0
+            voxel_grid.index_add_(dim=0,
+                                  index=xs[valid_indices] + ys[valid_indices] *
+                                  width + tis_long[valid_indices] * width * height,
+                                  source=vals_left[valid_indices])
 
-        valid_indices = (tis + 1) < num_bins
-        valid_indices &= tis >= 0
+            valid_indices = (tis + 1) < num_bins
+            valid_indices &= tis >= 0
 
-        voxel_grid.index_add_(dim=0,
-                                index=xs[valid_indices] + ys[valid_indices] * width +
-                                (tis_long[valid_indices] + 1) * width * height,
-                                source=vals_right[valid_indices])
+            voxel_grid.index_add_(dim=0,
+                                  index=xs[valid_indices] + ys[valid_indices] * width +
+                                  (tis_long[valid_indices] + 1) * width * height,
+                                  source=vals_right[valid_indices])
 
         voxel_grid = voxel_grid.view(num_bins, height, width)
